+++ src/multiplots.ts
@@ -1,7 +1,7 @@
 import {PlotData, Interactions} from './plot-data';
 import {Point2D} from './primitives';
 import { Attribute, PointFamily, check_package_version, Window, TypeOf, equals, Sort, export_to_txt, RubberBand } from './utils';
-import { PlotContour, PlotScatter, ParallelPlot, PrimitiveGroupContainer, newHistogram, Histogram, Frame } from './subplots';
+import { PlotContour, PlotScatter, ParallelPlot, PrimitiveGroupContainer, newHistogram, Frame } from './subplots';
 import { List, Shape, MyObject } from './toolbox';
 import { string_to_hex, string_to_rgb, rgb_to_string } from './color_conversion';
 
@@ -94,10 +94,11 @@ export class MultiplePlots {
               newObject = this.initialize_containers_dicts(newObject, association['associated_elements']);
               newObject = this.call_layout(newObject, association['attribute_names']);
             }
-          } else if (object_type_ === 'histogram') {
+          } else if (object_type_ === 'frame') {
             this.dataObjects[i]['elements'] = elements;
             newObject = new newHistogram(this.dataObjects[i], this.sizes[i]['width'], this.sizes[i]['height'], buttons_ON, this.initial_coords[i][0], this.initial_coords[i][1], canvas_id, true);
           } else {
+            console.log(this.dataObjects[i])
             throw new Error('MultiplePlots constructor : invalid object type');
           }
           this.initializeObjectContext(newObject);
@@ -858,9 +859,6 @@ export class MultiplePlots {
         } else if ((obj.type_ === 'parallelplot') && !List.isListOfEmptyList(obj.rubber_bands)) {
           bool = true;
           this.dep_selected_points_index = List.listIntersection(this.dep_selected_points_index, obj.pp_selected_index);
-        } else if ((obj.type_ === 'histogram') && !equals(obj.selected_point_index, [])) {
-          bool = true;
-          this.dep_selected_points_index = List.listIntersection(this.dep_selected_points_index, obj.selected_point_index);
         }
       }
 
@@ -899,7 +897,6 @@ export class MultiplePlots {
         let plot = this.objectList[i];
         if (plot instanceof PlotScatter) {Interactions.click_on_reset_action(plot)}
         else {plot.reset_scales()}
-        if (plot instanceof Histogram) {plot.reset_x_rubberband()}
       }
     }
 
@@ -916,8 +913,6 @@ export class MultiplePlots {
           otherPlot.reset_rubberbands();
         } else if (otherPlot instanceof Frame) {
           otherPlot.reset();
-        } else if (otherPlot instanceof Histogram) {
-          otherPlot.reset_x_rubberband();
         } else if (otherPlot instanceof PrimitiveGroupContainer) {
           otherPlot.reset_selection();
         }
@@ -1366,13 +1361,6 @@ export class MultiplePlots {
             subplot.axes[0].rubberBand.maxValue = rubberBand.maxValue;
             subplot.draw()
           })
-        } else if (subplot instanceof Histogram) {
-          rubberBandsInPlot.forEach((rubberBand) => {
-            let currentRubberIndex = rubberBandNames.indexOf(rubberBand.attributeName)
-            subplot.rubber_bands[0].updateFromOther(
-              rubberBand, subplot.axis_x_start, subplot.axis_x_end, false, currentPP.inverted_axis_list[currentRubberIndex]);
-          })
-          subplot.get_selected_keys();
         } else if (subplot instanceof PrimitiveGroupContainer) {
           subplot.selected_point_index = selectedIndices;
           if (selectedIndices.length == 0) {
@@ -1387,42 +1375,10 @@ export class MultiplePlots {
       })
     }
 
-    mouse_move_histogram_communication() {
-      let index = this.get_drawing_rubberbands_obj_index("histogram");
-      if (index === -1) return;
-      this.histogram_communication(index);
-    }
-
     mouse_move_frame_communication() {
       let index = this.get_drawing_rubberbands_obj_index("frame");
       if (index === -1) return;
       this.frame_communication(index);
-    }
-
-
-    histogram_communication(index) {
-      let histogram = this.objectList[index];
-      let primitive_indices = [];
-      for (let i=0; i<this.nbObjects; i++) {
-        let obj = this.objectList[i];
-        if (obj.type_ === 'scatterplot') {
-          MultiplotCom.histogram_to_scatter_communication(histogram, obj);
-        } else if (obj.type_ === 'parallelplot') {
-          MultiplotCom.histogram_to_pp_communication(histogram, obj);
-        } else if (obj.type_ === 'histogram') {
-          MultiplotCom.histogram_to_histogram_communication(histogram, obj);
-        } else if (obj.type_ === "primitivegroupcontainer") {
-          primitive_indices.push(i);
-        }
-      }
-      this.refresh_dep_selected_points_index();
-      this.refresh_selected_object_from_index();
-
-      for (let index of primitive_indices) {
-        let obj: any = this.objectList[index];
-        obj.selected_point_index = this.dep_selected_points_index;
-        obj.select_primitive_groups();
-      }
     }
 
     frame_communication(index) {
@@ -1906,7 +1862,6 @@ export class MultiplePlots {
               this.mouse_move_scatter_communication();
               this.mouse_move_pp_communication();
               this.mouse_move_frame_communication();
-              this.mouse_move_histogram_communication();
               this.redrawAllObjects();
             }
             this.refresh_selected_point_index();
@@ -1955,11 +1910,6 @@ export class MultiplePlots {
                   this.reset_selected_points_except([this.clickedPlotIndex]);
                   this.objectList[this.clickedPlotIndex].mouse_interaction(true);
                   this.pp_communication(this.objectList[this.clickedPlotIndex].rubber_bands, this.objectList[this.clickedPlotIndex]);
-                }
-              } else if (type_ === 'histogram') {
-                let obj: any = this.objectList[this.clickedPlotIndex];
-                if (obj.rubber_bands[0].length === 0) {
-                  this.reset_all_selected_points();
                 }
               } else if (type_ === "scatterplot") {
                 this.mouse_up_scatter_communication();
@@ -2101,53 +2051,6 @@ export class MultiplotCom {
       }
       Interactions.selection_window_action(plot_data);
     }
-
-    public static histogram_to_histogram_communication(histogram1, histogram2) {
-      if (histogram1.x_variable.name !== histogram2.x_variable.name) return;
-      histogram2.rubber_bands[0] = histogram1.rubber_bands[0];
-      histogram2.get_selected_keys();
-    }
-
-    public static histogram_to_pp_communication(histogram, parallel_plot) {
-      let index = -1;
-      for (let i=0; i<parallel_plot.axis_list.length; i++) {
-        if (histogram.x_variable.name === parallel_plot.axis_list[i].name) {
-          index = i;
-          break;
-        }
-      }
-      if (index === -1) return;
-      let x_variable = histogram.x_variable;
-      let x_rubberband = histogram.rubber_bands[0];
-      let axis_coord1 = parallel_plot.real_to_axis_coord(x_rubberband.minValue, x_variable.type_, x_variable.list,
-                                                          parallel_plot.inverted_axis_list[index]);
-      axis_coord1 = Math.max(Math.min(axis_coord1, 1), 0);
-      let axis_coord2 = parallel_plot.real_to_axis_coord(x_rubberband.maxValue, x_variable.type_, x_variable.list,
-                                                          parallel_plot.inverted_axis_list[index]);
-      axis_coord2 = Math.max(Math.min(axis_coord2, 1), 0);
-      parallel_plot.rubber_bands[index].axisMin = Math.min(axis_coord1, axis_coord2);
-      parallel_plot.rubber_bands[index].axisMax = Math.max(axis_coord1, axis_coord2);
-      parallel_plot.rubber_bands[index].minValue = x_rubberband.minValue;
-      parallel_plot.rubber_bands[index].maxValue = x_rubberband.maxValue;
-    }
-
-    public static histogram_to_scatter_communication(histogram, scatter) {
-      let scatter_x = scatter.plotObject.attribute_names[0];
-      let scatter_y = scatter.plotObject.attribute_names[1];
-      if (histogram.x_variable.name === scatter_x) {
-        scatter.perm_window_x = histogram.rubber_bands[0].minValue;
-        scatter.perm_window_y = -scatter.minY;
-        scatter.perm_window_w = histogram.rubber_bands[0].maxValue - histogram.rubber_bands[0].minValue;
-        scatter.perm_window_h = scatter.maxY - scatter.minY;
-      } else if (histogram.x_variable.name === scatter_y) {
-        scatter.perm_window_x = scatter.minX;
-        scatter.perm_window_y = histogram.rubber_bands[0].maxValue;
-        scatter.perm_window_w = scatter.maxX - scatter.minX;
-        scatter.perm_window_h = histogram.rubber_bands[0].maxValue - histogram.rubber_bands[0].minValue;
-      }
-      Interactions.selection_window_action(scatter)
-    }
-
 
     public static frame_to_frame_communication(histogram1, histogram2) {
       if (histogram1.axes[0].name !== histogram2.axes[0].name) return;