+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Add
 - Tooltip in Histograms
 
+### Refactor
+- Remove old Histogram and switch ScatterMatrix to Multiplot
+
 ## [0.14.1]
 ### Fix
 - plot_canvas function in python (needed width and height arguments) (in dev too)